requirements.txt
rembg==2.0.25
streamlit-1.14.1